22-01-13

+++ src/main/java/com/cslg/disk/common/BaseEntity.java
@@ -12,6 +12,7 @@ import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
+import java.sql.Time;
 import java.util.Date;
 
 
@@ -21,10 +22,8 @@ import java.util.Date;
 public abstract class BaseEntity {
 
     @Id
-//    @GeneratedValue(generator = "system-uuid")
-//    @GenericGenerator(name="system-uuid",strategy = "uuid")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
 
     // additional field
     @CreatedDate
